Controller Refactoring
