+++ packages/place-mark-api/app/core/endpoints/index.ts
@@ -1,2 +1,3 @@
 export * from "./controller.js";
 export * from "./decorator.js";
+export * from "./validation-result.js";
