+++ packages/place-mark-api/app/core/endpoints/validation-result.ts
@@ -1,0 +1,4 @@
+export interface IValidationResult {
+  property: string;
+  message: string;
+}
