+++ packages/place-mark-api/tests/unit-tests/controllers/api/user-api-controller.spec.ts
@@ -2,8 +2,7 @@ import { Server, ServerInjectResponse } from "@hapi/hapi";
 import { assert } from "chai";
 import { ContainerMock, testConfig } from "../test-setup.js";
 import { createServer$ } from "../../../../app/server.js";
-import { IPaginatedListRequest, IPaginatedListResponse, IUserReadOnlyDto, IUserRepository } from "../../../../app/core/index.js";
-import { IValidationResult } from "../../../../app/controllers/index.js";
+import { IPaginatedListRequest, IPaginatedListResponse, IUserReadOnlyDto, IUserRepository, IValidationResult } from "../../../../app/core/index.js";
 import { toQueryString, QueryParams } from "../../../utils.js";
 
 suite("UserApiController Unit-Tests", () => {
